Move tests, fix imports
