+++ tests/unit/doc.test.js
@@ -17,32 +17,15 @@ limitations under the License.
 
 import { JSDOM } from "jsdom";
 import {
-  domToString,
-  normaliseContent,
-  escapeCharacters,
-  cleanEscapeCharacters,
-  fixBrokenHTML,
-  getNodeWithoutClasses,
   getDocument,
-  getDocumentNodeFromString,
-  createHTML,
   createFeedXML,
-  dumpNode,
-  getDoctype,
-  getDocumentContentNode,
   createActivityHTML,
-  getClosestSectionNode,
-  removeSelectorFromNode,
-  getNodeLanguage,
   addMessageToLog,
-  selectArticleNode,
   insertDocumentLevelHTML,
   setDate,
   createDateHTML,
-  getRDFaPrefixHTML,
   getDocumentStatusHTML,
   getGraphData,
-  parseMarkdown,
   serializeTableToText,
   createLanguageHTML,
   createInboxHTML,
@@ -56,8 +39,8 @@ import {
 } from "../../src/doc";
 import Config from "../../src/config";
 import MockGrapoi from "../utils/mockGrapoi";
-import { domSanitize, generateAttributeId } from "../../src/util";
-import rdf from 'rdf-ext';
+import rdf from "rdf-ext";
+import { vi } from "vitest";
 
 vi.mock(import("../../src/util"), async (importOriginal) => {
   const actual = await importOriginal();
@@ -140,86 +123,6 @@ beforeAll(() => {
   });
 });
 
-describe("domToString", () => {
-  it("serializes a clean DOM", () => {
-    const htmlContent = `<head><title>Test</title></head><body><p>asdf</p></body>`;
-
-    const node = window.document.createElement("html");
-    node.innerHTML = htmlContent;
-
-    const result = domToString(node);
-
-    const expected =
-      `<html>
-${htmlContent}
-</html>`
-
-    expect(result).toBe(expected);
-  });
-});
-
-describe("getDocumentNodeFromString", () => {
-  it("parses HTML string to document node", () => {
-    const htmlContent = `<html><head><title>Test</title></head><body><p>asdf</p></body></html>`;
-    const expected = `<html>
-<head><title>Test</title></head><body><p>asdf</p></body>
-</html>`;
-    let result = getDocumentNodeFromString(htmlContent);
-    result = domToString(result.documentElement);
-
-    expect(result).toBe(expected);
-  })
-})
-
-describe('normaliseContent', () => {
-  it.only('normalizes content after potential prosemirror schema changes', () => {
-    const htmlContent = `<html><head><title>Test</title></head><body><li><p>asdf</p></li></body></html>`;
-    const dom = new JSDOM(htmlContent);
-    const node = dom.window.document.documentElement.cloneNode(true);
-    const expected = `<html><head><title>Test</title></head><body><li>asdf</li></body></html>`;
-    const result = normaliseContent(node)
-    expect(result).not.toBeUndefined();
-    const div = dom.window.document.createElement("div");
-    div.appendChild(result);
-    expect(div.firstChild.outerHTML).toBe(expected);
-  });
-})
-
-describe("escapeCharacters", () => {
-  it("escapes special characters correctly", () => {
-    const input = "<div>&\"'</div>";
-    const expectedOutput = "&lt;div&gt;&amp;&quot;&apos;&lt;/div&gt;";
-    expect(escapeCharacters(input)).toBe(expectedOutput);
-  });
-});
-
-describe("cleanEscapeCharacters", () => {
-  it("cleans double escaped characters correctly", () => {
-    const input = "&amp;lt;&amp;gt;&amp;apos;&amp;quot;&amp;amp;";
-    const expectedOutput = "&lt;&gt;&apos;&quot;&amp;";
-    expect(cleanEscapeCharacters(input)).toBe(expectedOutput);
-  });
-});
-
-describe("fixBrokenHTML", () => {
-  it("fixes img", () => {
-    const input = '<img src="image.jpg"></img>';
-    const expectedOutput = '<img src="image.jpg"/>';
-    expect(fixBrokenHTML(input)).toBe(expectedOutput);
-  });
-});
-
-describe("getNodeWithoutClasses", () => {
-  it("removes specified classes from node", () => {
-    const node = dom.window.document.createElement("div");
-    node.innerHTML =
-      '<span class="remove-me">Text</span><span class="keep-me">Text</span>';
-    const resultNode = getNodeWithoutClasses(node, "remove-me");
-    expect(resultNode.querySelector(".remove-me")).toBeNull();
-    expect(resultNode.querySelector(".keep-me")).not.toBeNull();
-  });
-});
-
 describe("getDocument", () => {
   it("returns document as string", () => {
     const result = getDocument(dom.window.document.documentElement);
@@ -227,28 +130,6 @@ describe("getDocument", () => {
     expect(result).toContain(
       '<html lang="en" xml:lang="en" xmlns="http://www.w3.org/1999/xhtml">'
     );
-  });
-});
-
-describe("getDocumentNodeFromString", () => {
-  it("parses string to document node", () => {
-    const resultNode = getDocumentNodeFromString(htmlContent);
-    expect(resultNode.querySelector("title")).not.toBeNull();
-  });
-});
-
-describe("createHTML", () => {
-  it("creates HTML string with given title and main content", () => {
-    const title = "title";
-    const main = "<p>content</p>";
-    const result = createHTML(title, main);
-
-    const normalizedResult = result.replace(/\s+/g, "").trim();
-    const expectedTitle = "<title>title</title>";
-    const expectedMain = "<main><p>content</p></main>";
-
-    expect(normalizedResult).toContain(expectedTitle);
-    expect(normalizedResult).toContain(expectedMain);
   });
 });
 
@@ -364,122 +245,6 @@ describe("createFeedXML", () => {
   });
 });
 
-describe("dumpNode", () => {
-  let options, skipAttributes, voidElements, noEsc;
-
-  beforeEach(() => {
-    options = {
-      skipNodeWithId: [],
-      classWithChildText: {
-        class: "child-class",
-        element: "span",
-      },
-      skipNodeWithClass: "",
-      replaceClassItemWith: {
-        source: ["old-class"],
-        target: "new-class",
-      },
-      sortAttributes: true,
-      skipEscapingDataBlockTypes: [],
-    };
-    skipAttributes = [];
-    voidElements = [
-      "area",
-      "base",
-      "br",
-      "col",
-      "embed",
-      "hr",
-      "img",
-      "input",
-      "keygen",
-      "link",
-      "meta",
-      "param",
-      "source",
-      "track",
-      "wbr",
-    ];
-    noEsc = [];
-  });
-
-  it("should return empty string for non-nodes", () => {
-    const s = "Sample non-node";
-    expect(dumpNode(s, options, skipAttributes, voidElements, noEsc)).toBe("");
-  });
-
-  it("should handle element nodes correctly", () => {
-    const divNode = document.createElement("div");
-    divNode.setAttribute("id", "test");
-    divNode.setAttribute("class", "test-class");
-
-    const spanNode = document.createElement("span");
-    spanNode.textContent = "Hello";
-    divNode.appendChild(spanNode);
-
-    const result = dumpNode(
-      divNode,
-      options,
-      skipAttributes,
-      voidElements,
-      noEsc
-    );
-    expect(result).toBe(
-      '<div class="test-class" id="test"><span>Hello</span></div>'
-    );
-  });
-
-  it("should skip nodes with specific IDs", () => {
-    options.skipNodeWithId.push("test");
-    const divNode = document.createElement("div");
-    divNode.setAttribute("id", "test");
-
-    const result = dumpNode(
-      divNode,
-      options,
-      skipAttributes,
-      voidElements,
-      noEsc
-    );
-    expect(result).toBe("");
-  });
-});
-
-describe("getDoctype", () => {
-  it("should return correct DOCTYPE string", () => {
-    document.implementation.createHTMLDocument();
-    expect(getDoctype()).toBe("<!DOCTYPE html>");
-  });
-});
-
-describe("getDocumentContentNode", () => {
-  describe("getDocumentContentNode", () => {
-    it("should return body for HTMLDocument", () => {
-      const newDom = new JSDOM(`<!DOCTYPE html><html><body></body></html>`, {
-        url: "https://example.com/",
-      });
-      global.window = newDom.window;
-      global.document = newDom.window.document;
-      global.Document = newDom.window.Document;
-      global.HTMLElement = newDom.window.HTMLElement;
-      expect(getDocumentContentNode(document)).toBe(document.body);
-    });
-
-    it("should return first child for DocumentFragment", () => {
-      const document = dom.window.document;
-
-      const fragment = new DocumentFragment();
-      const divNode = document.createElement("div");
-      fragment.appendChild(divNode);
-      expect(getDocumentContentNode(fragment)).toBe(divNode);
-    });
-  });
-
-  it("should return undefined for unknown document types", () => {
-    const unknownNode = {};
-    expect(getDocumentContentNode(unknownNode)).toBeUndefined();
-  });
-});
 describe("createActivityHTML", () => {
   it("createActivityHTML returns correct HTML structure", () => {
     const o = {
@@ -505,52 +270,6 @@ describe("createActivityHTML", () => {
   });
 });
 
-describe("getClosestSectionNode", () => {
-  it(" returns the closest section node", () => {
-    document.body.innerHTML = `
-    <div>
-      <article>
-        <section id="section1">
-          <div id="testNode"></div>
-        </section>
-      </article>
-    </div>`;
-
-    const node = document.getElementById("testNode");
-    const result = getClosestSectionNode(node);
-
-    expect(result.tagName.toLowerCase()).toBe("section");
-  });
-});
-
-it("removeSelectorFromNode removes the specified selector from node", () => {
-  document.body.innerHTML = `
-    <div>
-      <p class="removeMe">Text1</p>
-      <p>Text2</p>
-    </div>`;
-
-  const node = document.querySelector("div");
-  const clone = removeSelectorFromNode(node, ".removeMe");
-
-  expect(clone.querySelector(".removeMe")).toBeNull();
-  expect(clone.querySelectorAll("p").length).toBe(1);
-});
-
-describe("getNodeLanguage", () => {
-  it("getNodeLanguage returns the correct language attribute", () => {
-    document.body.innerHTML = `
-    <div lang="en">
-      <p lang="fr" id="testNode">Text</p>
-    </div>`;
-
-    const node = document.getElementById("testNode");
-    const result = getNodeLanguage(node);
-
-    expect(result).toBe("fr");
-  });
-});
-
 describe("addMessageToLog", () => {
   it("addMessageToLog adds a message with dateTime to Config.MessageLog", () => {
     const Config = { MessageLog: [] };
@@ -563,28 +282,7 @@ describe("addMessageToLog", () => {
     expect(Config.MessageLog[0].content).toBe("New message");
   });
 });
-describe("selectArticleNode", () => {
-  it("should select the last matching article node", () => {
-    const document = dom.window.document;
-    const result = selectArticleNode(document.body);
-    expect(result.nodeName).toBe("ARTICLE");
-  });
 
-  it("should return default content node when no article node is found", () => {
-    const dom = new JSDOM(`<!DOCTYPE html><html><body></body></html>`, {
-      url: "https://example.com/",
-    });
-    const document = dom.window.document;
-
-    global.window = dom.window;
-    global.document = dom.window.document;
-    global.Document = dom.window.Document;
-    global.HTMLElement = dom.window.HTMLElement;
-
-    const result = selectArticleNode(document.body);
-    expect(result).toBe(document.body);
-  });
-});
 
 describe("insertDocumentLevelHTML", () => {
   document = dom.window.document;
@@ -686,25 +384,6 @@ describe("createDateHTML", () => {
     expect(result).toContain(
       '<time content="2024-10-15T00:00:00.000Z" datatype="xsd:dateTime" datetime="2024-10-15T00:00:00.000Z" property="schema:dateCreated">2024-10-15</time>'
     );
-  });
-});
-
-describe("getRDFaPrefixHTML", () => {
-  it("should return formatted prefix HTML", () => {
-    const prefixes = {
-      foaf: "http://xmlns.com/foaf/0.1/",
-      rdf: "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
-    };
-    const result = getRDFaPrefixHTML(prefixes);
-    expect(result).toContain("foaf: http://xmlns.com/foaf/0.1/");
-    expect(result).toContain(
-      "rdf: http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-    );
-  });
-
-  it("should handle empty prefix object", () => {
-    const result = getRDFaPrefixHTML({});
-    expect(result).toBe("");
   });
 });
 
@@ -858,32 +537,6 @@ describe("serializeTableToText", () => {
 
     expect(result).toContain("Header 1");
     expect(result).not.toContain("Data");
-  });
-});
-
-describe("parseMarkdown", () => {
-  it("should parse markdown into HTML without creating a document", () => {
-    const markdown = "# Test Markdown";
-    const result = parseMarkdown(markdown);
-
-    expect(result).toContain("<h1>Test Markdown</h1>");
-  });
-
-  it("should parse markdown and create an article element when createDocument is true", () => {
-    const markdown = "# Test Markdown";
-    const options = { createDocument: true };
-    const result = parseMarkdown(markdown, options);
-
-    expect(result).toContain("<article>");
-    expect(result).toContain("<h1>Test Markdown</h1>");
-    expect(result).toContain("</article>");
-  });
-
-  it("should parse markdown with HTML correctly", () => {
-    const markdown = "Some <b>bold</b> text";
-    const result = parseMarkdown(markdown);
-
-    expect(result).toContain("<b>bold</b>");
   });
 });
 
@@ -1088,135 +741,217 @@ describe("createRDFaHTML", () => {
 
 function getResourceInfoODRLPolicies(s) {
   var info = {};
-  info['odrl'] = {};
+  info["odrl"] = {};
 
   var policy = s.out(ns.odrl.hasPolicy);
 
-  policy.values.forEach(policyIRI => {
-    info['odrl'][policyIRI] = {};
+  policy.values.forEach((policyIRI) => {
+    info["odrl"][policyIRI] = {};
 
     var policyGraph = s.node(rdf.namedNode(policyIRI));
     var policyTypes = policyGraph.out(ns.rdf.type).values;
 
-    info['odrl'][policyIRI]['rdftype'] = policyTypes;
+    info["odrl"][policyIRI]["rdftype"] = policyTypes;
 
-    policyTypes.forEach(pT => {
+    policyTypes.forEach((pT) => {
       if (pT == ns.odrl.Offer.value) {
         var permissions = policyGraph.out(ns.odrl.permission).values;
 
-        permissions.forEach(permissionIRI => {
-          info['odrl'][policyIRI]['permission'] = {};
-          info['odrl'][policyIRI]['permission'][permissionIRI] = {};
+        permissions.forEach((permissionIRI) => {
+          info["odrl"][policyIRI]["permission"] = {};
+          info["odrl"][policyIRI]["permission"][permissionIRI] = {};
 
           var permissionGraph = s.node(rdf.namedNode(permissionIRI));
 
           var permissionAssigner = permissionGraph.out(ns.odrl.assigner).values;
-          info['odrl'][policyIRI]['permission'][permissionIRI]['action'] = info['odrl']['permissionAssigner'] = permissionAssigner;
+          info["odrl"][policyIRI]["permission"][permissionIRI]["action"] = info[
+            "odrl"
+          ]["permissionAssigner"] = permissionAssigner;
 
           var permissionActions = permissionGraph.out(ns.odrl.action).values;
-          info['odrl'][policyIRI]['permission'][permissionIRI]['action'] = info['odrl']['permissionActions'] = permissionActions;
+          info["odrl"][policyIRI]["permission"][permissionIRI]["action"] = info[
+            "odrl"
+          ]["permissionActions"] = permissionActions;
         });
       }
 
       if (pT == ns.odrl.Agreement.value) {
         var prohibition = policyGraph.out(ns.odrl.prohibition).values;
 
-        prohibition.forEach(prohibitionIRI => {
-          info['odrl'][policyIRI]['prohibition'] = {};
-          info['odrl'][policyIRI]['prohibition'][prohibitionIRI] = {};
+        prohibition.forEach((prohibitionIRI) => {
+          info["odrl"][policyIRI]["prohibition"] = {};
+          info["odrl"][policyIRI]["prohibition"][prohibitionIRI] = {};
 
           var prohibitionGraph = s.node(rdf.namedNode(prohibitionIRI));
 
-          var prohibitionAssigner = prohibitionGraph.out(ns.odrl.assigner).values;
-          info['odrl'][policyIRI]['prohibition'][prohibitionIRI]['action'] = info['odrl']['prohibitionAssigner'] = prohibitionAssigner;
+          var prohibitionAssigner = prohibitionGraph.out(
+            ns.odrl.assigner
+          ).values;
+          info["odrl"][policyIRI]["prohibition"][prohibitionIRI]["action"] =
+            info["odrl"]["prohibitionAssigner"] = prohibitionAssigner;
 
-          var prohibitionAssignee = prohibitionGraph.out(ns.odrl.assignee).values;
-          info['odrl'][policyIRI]['prohibition'][prohibitionIRI]['action'] = info['odrl']['prohibitionAssignee'] = prohibitionAssignee;
+          var prohibitionAssignee = prohibitionGraph.out(
+            ns.odrl.assignee
+          ).values;
+          info["odrl"][policyIRI]["prohibition"][prohibitionIRI]["action"] =
+            info["odrl"]["prohibitionAssignee"] = prohibitionAssignee;
 
           var prohibitionActions = prohibitionGraph.out(ns.odrl.action).values;
-          info['odrl'][policyIRI]['prohibition'][prohibitionIRI]['action'] = info['odrl']['prohibitionActions'] = prohibitionActions;
+          info["odrl"][policyIRI]["prohibition"][prohibitionIRI]["action"] =
+            info["odrl"]["prohibitionActions"] = prohibitionActions;
         });
       }
     });
   });
 
-  return info['odrl'];
+  return info["odrl"];
 }
 
-describe('getResourceInfoODRLPolicies with MockGrapoi', () => {
-  it('processes Offer policies with permissions correctly', () => {
-    const policyIRI = 'http://example.org/policy/1';
-    const permissionIRI = 'http://example.org/permission/1';
+describe("getResourceInfoODRLPolicies with MockGrapoi", () => {
+  it("processes Offer policies with permissions correctly", () => {
+    const policyIRI = "http://example.org/policy/1";
+    const permissionIRI = "http://example.org/permission/1";
 
     const s = new MockGrapoi([
-      { subject: { value: 'root' }, predicate: ns.odrl.hasPolicy, object: { value: policyIRI } },
-      { subject: { value: policyIRI }, predicate: ns.rdf.type, object: { value: ns.odrl.Offer.value } },
-      { subject: { value: policyIRI }, predicate: ns.odrl.permission, object: { value: permissionIRI } },
+      {
+        subject: { value: "root" },
+        predicate: ns.odrl.hasPolicy,
+        object: { value: policyIRI },
+      },
+      {
+        subject: { value: policyIRI },
+        predicate: ns.rdf.type,
+        object: { value: ns.odrl.Offer.value },
+      },
+      {
+        subject: { value: policyIRI },
+        predicate: ns.odrl.permission,
+        object: { value: permissionIRI },
+      },
 
-      { subject: { value: permissionIRI }, predicate: ns.odrl.assigner, object: { value: 'assigner1' } },
-      { subject: { value: permissionIRI }, predicate: ns.odrl.action, object: { value: 'action1' } },
-      { subject: { value: permissionIRI }, predicate: ns.odrl.action, object: { value: 'action2' } },
+      {
+        subject: { value: permissionIRI },
+        predicate: ns.odrl.assigner,
+        object: { value: "assigner1" },
+      },
+      {
+        subject: { value: permissionIRI },
+        predicate: ns.odrl.action,
+        object: { value: "action1" },
+      },
+      {
+        subject: { value: permissionIRI },
+        predicate: ns.odrl.action,
+        object: { value: "action2" },
+      },
     ]);
 
-    s.node('root');
+    s.node("root");
 
     const info = getResourceInfoODRLPolicies(s);
 
     expect(info[policyIRI].rdftype).toEqual([ns.odrl.Offer.value]);
     expect(Object.keys(info[policyIRI].permission)).toContain(permissionIRI);
-    expect(info[policyIRI].permission[permissionIRI].action).toEqual(['action1', 'action2']);
-    expect(info.permissionAssigner).toEqual(['assigner1']);
-    expect(info.permissionActions).toEqual(['action1', 'action2']);
+    expect(info[policyIRI].permission[permissionIRI].action).toEqual([
+      "action1",
+      "action2",
+    ]);
+    expect(info.permissionAssigner).toEqual(["assigner1"]);
+    expect(info.permissionActions).toEqual(["action1", "action2"]);
   });
 
-  it('processes Agreement policies with prohibitions correctly', () => {
-    const policyIRI = 'http://example.org/policy/2';
-    const prohibitionIRI = 'http://example.org/prohibition/1';
+  it("processes Agreement policies with prohibitions correctly", () => {
+    const policyIRI = "http://example.org/policy/2";
+    const prohibitionIRI = "http://example.org/prohibition/1";
 
     const s = new MockGrapoi([
-      { subject: { value: 'root' }, predicate: ns.odrl.hasPolicy, object: { value: policyIRI } },
-      { subject: { value: policyIRI }, predicate: ns.rdf.type, object: { value: ns.odrl.Agreement.value } },
-      { subject: { value: policyIRI }, predicate: ns.odrl.prohibition, object: { value: prohibitionIRI } },
+      {
+        subject: { value: "root" },
+        predicate: ns.odrl.hasPolicy,
+        object: { value: policyIRI },
+      },
+      {
+        subject: { value: policyIRI },
+        predicate: ns.rdf.type,
+        object: { value: ns.odrl.Agreement.value },
+      },
+      {
+        subject: { value: policyIRI },
+        predicate: ns.odrl.prohibition,
+        object: { value: prohibitionIRI },
+      },
 
-      { subject: { value: prohibitionIRI }, predicate: ns.odrl.assigner, object: { value: 'prohibitionAssigner1' } },
-      { subject: { value: prohibitionIRI }, predicate: ns.odrl.assignee, object: { value: 'prohibitionAssignee1' } },
-      { subject: { value: prohibitionIRI }, predicate: ns.odrl.action, object: { value: 'prohibitionAction1' } },
-      { subject: { value: prohibitionIRI }, predicate: ns.odrl.action, object: { value: 'prohibitionAction2' } },
+      {
+        subject: { value: prohibitionIRI },
+        predicate: ns.odrl.assigner,
+        object: { value: "prohibitionAssigner1" },
+      },
+      {
+        subject: { value: prohibitionIRI },
+        predicate: ns.odrl.assignee,
+        object: { value: "prohibitionAssignee1" },
+      },
+      {
+        subject: { value: prohibitionIRI },
+        predicate: ns.odrl.action,
+        object: { value: "prohibitionAction1" },
+      },
+      {
+        subject: { value: prohibitionIRI },
+        predicate: ns.odrl.action,
+        object: { value: "prohibitionAction2" },
+      },
     ]);
 
-    s.node('root');
+    s.node("root");
 
     const info = getResourceInfoODRLPolicies(s);
 
     expect(info[policyIRI].rdftype).toEqual([ns.odrl.Agreement.value]);
     expect(Object.keys(info[policyIRI].prohibition)).toContain(prohibitionIRI);
-    expect(info[policyIRI].prohibition[prohibitionIRI].action).toEqual(['prohibitionAction1', 'prohibitionAction2']);
-    expect(info.prohibitionAssigner).toEqual(['prohibitionAssigner1']);
-    expect(info.prohibitionAssignee).toEqual(['prohibitionAssignee1']);
-    expect(info.prohibitionActions).toEqual(['prohibitionAction1', 'prohibitionAction2']);
+    expect(info[policyIRI].prohibition[prohibitionIRI].action).toEqual([
+      "prohibitionAction1",
+      "prohibitionAction2",
+    ]);
+    expect(info.prohibitionAssigner).toEqual(["prohibitionAssigner1"]);
+    expect(info.prohibitionAssignee).toEqual(["prohibitionAssignee1"]);
+    expect(info.prohibitionActions).toEqual([
+      "prohibitionAction1",
+      "prohibitionAction2",
+    ]);
   });
 
-  it('returns empty object if no policies', () => {
+  it("returns empty object if no policies", () => {
     const s = new MockGrapoi([]);
-    s.node('root');
+    s.node("root");
 
     const info = getResourceInfoODRLPolicies(s);
 
     expect(info).toEqual({});
   });
-  it('skips unknown policy types', () => {
-    const policyIRI = 'http://example.org/policy/3';
-  
+  it("skips unknown policy types", () => {
+    const policyIRI = "http://example.org/policy/3";
+
     const s = new MockGrapoi([
-      { subject: { value: 'root' }, predicate: ns.odrl.hasPolicy, object: { value: policyIRI } },
-      { subject: { value: policyIRI }, predicate: ns.rdf.type, object: { value: 'http://example.org/UnknownPolicyType' } },
+      {
+        subject: { value: "root" },
+        predicate: ns.odrl.hasPolicy,
+        object: { value: policyIRI },
+      },
+      {
+        subject: { value: policyIRI },
+        predicate: ns.rdf.type,
+        object: { value: "http://example.org/UnknownPolicyType" },
+      },
     ]);
-  
-    s.node('root');
-  
+
+    s.node("root");
+
     const info = getResourceInfoODRLPolicies(s);
-  
-    expect(info[policyIRI].rdftype).toEqual(['http://example.org/UnknownPolicyType']);
+
+    expect(info[policyIRI].rdftype).toEqual([
+      "http://example.org/UnknownPolicyType",
+    ]);
     expect(info[policyIRI].permission).toBeUndefined();
     expect(info[policyIRI].prohibition).toBeUndefined();
   });
