+++ tests/unit/ui/buttons.test.js
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { getButtonHTML, buttonIcons } from 'src/ui/button-icons'; 
+import { getButtonHTML, buttonIcons } from 'src/ui/buttons.js'; 
 
 describe('getButtonHTML', () => {
   beforeEach(() => {
